Make it work actually. Need to refactor

+++ reactts-listapp/src/myapproach/model/model.ts
@@ -1,5 +1,5 @@
 import { api } from "../db/Database";
-import { v4 } from "uuid";
+import { v4 as uuid } from "uuid";
 
 export type Item = {
   id: string;
@@ -9,21 +9,17 @@ export type Item = {
   updatedAt: number;
 };
 
-// type CRUDResult = {
-//   id: string;
-//   message: string;
-//   operatedAt: number;
-//   operationType: "create" | "remove" | "update";
-// };
-
-export class ItemList {
+export class Model {
   private list: Item[];
+  private onreadyCallbacks: (() => void)[] = [];
+
   constructor(list: Item[] = []) {
     this.list = list;
   }
+
   newItem(label: string): Item {
     return {
-      id: v4().slice(0, 8),
+      id: uuid().slice(0, 8),
       label,
       checked: false,
       createdAt: Date.now(),
@@ -47,6 +43,7 @@ export class ItemList {
     const item = this.list.find((item) => item.id === itemId);
     return item ? { ...item } : undefined;
   };
+
   public readItems = (itemId?: string[] | undefined): Item[] | undefined => {
     api.read(itemId);
     if (itemId === undefined) {
@@ -64,15 +61,19 @@ export class ItemList {
     this.list = this.list.filter((item) => item.id !== itemId);
     api.delete(itemId);
   };
+
   public removeItems = (itemId: string[]): void => {
-    this.list = this.list.filter((item) =>
-      itemId.some((iid) => iid !== item.id),
-    );
-    api.delete(itemId);
+    if(itemId.length) {
+      console.log(this.list);
+      this.list = this.list.filter((item) => !itemId.some((id) => id === item.id));
+
+      console.log(this.list);
+      api.delete(itemId);
+    }
   };
 
   public clearItems = (): void => {
-    const itemsToClear = this.list.filter(item => item.checked).map(item => item.id);
+    const itemsToClear = this.list.filter(({ checked }) => checked).map(item => item.id);
     console.log(itemsToClear);
     this.removeItems(itemsToClear)
   }
@@ -84,29 +85,26 @@ export class ItemList {
       api.update(itemId, item);
     }
   };
-  public setItemCheck = (itemId: string, checked: boolean): void => {
+
+  public setItemChecked = (itemId: string, checked: boolean): void => {
     const item = this.list.find((item) => item.id === itemId);
     if (item) {
-      console.log(`${item.label} is ${!checked ? "not " : ""}done`);
+      //console.log(`${item.label} is ${!checked ? "not " : ""}done`);
       item.checked = checked;
       api.update(itemId, item);
     }
   };
 
-
-  private onreadyCallbacks: (() => void)[] = [];
   public onready = (callback: () => void): void => {
     this.onreadyCallbacks.push(callback);
   }
+
   public init = (): void => {
     api.read(undefined)
       .then((listt) => {
         this.list = listt?.data as Item[];
         this.onreadyCallbacks.forEach(callback => callback());
-        //console.log(this.list);
     });
   };
 }
-
-export const itemList = new ItemList();
